requirements.txt
playwright>=1.44.0
python-telegram-bot>=20.7
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
pytz>=2024.1
